component 옛날파일 삭제. 새 파일에 변경사항 반영
